[tests] Change some `runNpmInstall()` unit test to mocks
